Add files via upload
Mission 4 "Pluto chase". Just a rough map. I only set down a few props to give an idea on what could be used to spice up the map. The small outpost is the only ISDF buildings that will ever be present. Meant to be more of a resupply point. And the simple pluto terrain really gets boring at times so I had to break away a bit. Drew a path where the chase "can" take place it's rather big so lots of room to rework things.